Add styles to Navbar
