+++ src/App.js
@@ -1,6 +1,6 @@
 import './App.css';
 import { Routes, Route } from 'react-router-dom';
-import Header from './components/Header';
+import Navbar from './components/Navbar';
 import Dragon from './components/pages/Dragon';
 import Mission from './components/pages/Mission';
 import MyProfile from './components/pages/MyProfile';
@@ -9,12 +9,12 @@ import Rocket from './components/pages/Rocket';
 function App() {
   return (
     <div className="App">
-      <Header />
+      <Navbar />
       <Routes>
-        <Route path="rocket" element={<Rocket />} exact="true" />
-        <Route path="/missons" element={<Mission />} />
-        <Route path="/dragons" element={<Dragon />} />
-        <Route path="/profile" element={<MyProfile />} />
+        <Route path="rocket" element={<Rocket />} exact />
+        <Route path="/missons" element={<Mission />} exact />
+        <Route path="/dragons" element={<Dragon />} exact />
+        <Route path="/profile" element={<MyProfile />} exact />
       </Routes>
     </div>
   );
