+++ src/components/Navbar.css
@@ -1,0 +1,70 @@
+.nav_bar {
+  align-self: center;
+  padding: 24px 4%;
+  background-color: rgb(255, 255, 255);
+  display: flex;
+  align-items: center;
+  gap: 40px;
+}
+
+.logo {
+  display: flex;
+  flex-direction: row;
+}
+
+.title {
+  align-self: center;
+  font-family: Verdana, Geneva, Tahoma, sans-serif;
+  font-size: 30px;
+  font-weight: bold;
+  color: #202325;
+  cursor: pointer;
+  padding-left: 10px;
+}
+
+.nav-links {
+  display: flex;
+  justify-content: flex-end;
+  flex-grow: 1;
+  list-style-type: none;
+  gap: 1px;
+}
+
+.nav-link {
+  font-family: Verdana, Geneva, Tahoma, sans-serif;
+  font-size: 14px;
+  color: #121212;
+  letter-spacing: 1.9px;
+}
+
+.logo-image {
+  height: 60px;
+  width: 60px;
+}
+
+ul {
+  list-style: none;
+}
+
+li {
+  display: inline-block;
+  padding: 20px 0 20px;
+  vertical-align: middle;
+}
+
+.nav_bar ul li {
+  padding: 0;
+}
+
+a {
+  position: relative;
+  display: block;
+  padding: 5px 0;
+  margin: 0 12px;
+  letter-spacing: 1px;
+  font-size: 12px;
+  line-height: 16px;
+  font-weight: 900;
+  color: #222525;
+  text-decoration: none;
+}
