+++ src/components/pages/Mission.js
@@ -1,5 +1,7 @@
 const Mission = () => (
-  <div />
+  <div>
+    Missions
+  </div>
 );
 
 export default Mission;
